requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
PyPDF2
scikit-learn
Jinja2